refactor(upgrade-signal): split config and runtime modules
Amp-Thread-ID: https://ampcode.com/threads/T-019ebd47-648f-759e-8430-f9b506a603b9
